--- a/scipy/fft/_helper.py
+++ b/scipy/fft/_helper.py
@@ -62,6 +62,8 @@ def next_fast_len(target, real=False):
     pass
 
 
+# Directly wrap the c-function good_size but take the docstring etc. from the
+# next_fast_len function above
 next_fast_len = update_wrapper(lru_cache()(_helper.good_size), next_fast_len)
 next_fast_len.__wrapped__ = _helper.good_size