--- a/scipy/fftpack/helper.py
+++ b/scipy/fftpack/helper.py
@@ -95,7 +95,7 @@ def next_fast_len(target):
 
     """
     # Real transforms use regular sizes so this is backwards compatible
-    return _helper.next_fast_len(target, 'R2C')
+    return _helper.next_fast_len(target, True)
 
 
 def _good_shape(x, shape, axes):