#### Reference issue
Closes #10804, closes #10589

#### What does this implement/fix?
This supercedes #10804, extending it to also speed up `next_fast_len.__wrapped__`. The python overhead is brought down from ~550 ns to ~85 ns by binding the function directly in the c-api and replacing the `kind` argument with the `real` argument used in the C++ binding version.

#### Additional information
Here are the improved timings I see:
```
In [1]: from scipy.fft import next_fast_len
In [2]: %timeit next_fast_len.__wrapped__(12)
86.8 ns ± 3.78 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
In [3]: %timeit next_fast_len.__wrapped__(997)
222 ns ± 3.43 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
In [4]: %timeit next_fast_len.__wrapped__(1000)
135 ns ± 0.593 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
In [5]: %timeit next_fast_len.__wrapped__(16381)
415 ns ± 10.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

cc @larsoner @mreineck