@larsoner 2019-09-11
Travis failure looks legitimate:
```
=============
scipy.fftpack
=============
scipy.fftpack.next_fast_len
---------------------------
File "build/testenv/lib/python3.6/site-packages/scipy/fftpack/helper.py", line 87, in next_fast_len
Failed example:
    fftpack.helper.next_fast_len(min_len)
Exception raised:
    Traceback (most recent call last):
      File "/opt/python/3.6.7/lib/python3.6/doctest.py", line 1330, in __run
        compileflags, 1), test.globs)
      File "<doctest next_fast_len[4]>", line 1, in <module>
        fftpack.helper.next_fast_len(min_len)
      File "/home/travis/build/scipy/scipy/build/testenv/lib/python3.6/site-packages/scipy/fftpack/helper.py", line 98, in next_fast_len
        return _helper.next_fast_len(target, True)
    AttributeError: module 'scipy.fft._pocketfft.helper' has no attribute 'next_fast_len'
```